backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
fastapi==0.78.0
Flask==2.1.2
googleapis-common-protos==1.56.2
grpcio==1.46.3
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-asgi==0.29b0
opentelemetry-instrumentation-fastapi==0.29b0
opentelemetry-instrumentation-flask==0.29b0
opentelemetry-instrumentation-requests==0.29b0
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
prometheus-client==0.14.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pydantic==1.9.1
requests==2.28.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
thrift==0.16.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
